Project-1-ChatBot/requirements.txt
langchain
langchain_groq
streamlit
groq
python-dotenv
fastapi
uvicorn
langserve
sse_starlette
langchain_community